Update gemspec/pin to match Chef 14
Pin highline and mixlib-shellout to the versions in chef 14.

Without this pin we pull in highline 2.0 as part of chef-server
omnibus builds. The presence of highline 2.0 triggers some issues with
a number of tools, including chef-server's dvm tool because chef 14.2
pins highline to ~> 1.6.

This is strictly a short term fix while we figure out the long term
solution, probably involving appbundling more pieces of chef server to
make them well behaved.

Signed-off-by: Mark Anderson <mark@chef.io>